Refactor SequenceDiagram
