+++ test/uml/diagrams/sequence/sequencediagram/SequenceDiagramsInvalidNamesTest.java
@@ -1,4 +1,4 @@
-package uml.diagrams.sequence;
+package uml.diagrams.sequence.sequencediagram;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -9,6 +9,7 @@ import java.util.Collection;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import uml.diagrams.sequence.BaseElement;
 import uml.diagrams.sequence.exceptions.SequenceDiagramRuleException;
 
 
@@ -25,7 +26,9 @@ public class SequenceDiagramsInvalidNamesTest {
     void testErrorCreateSequenceDiagrams(String name, Boolean guardCondition) {
         SequenceDiagramRuleException exception = assertThrows(
                 SequenceDiagramRuleException.class,
-                () -> new SequenceDiagrams(name, guardCondition));
+                () -> new SequenceDiagram(name, guardCondition));
+        
+        String message = exception.getMessage();
         
         assertEquals(exception.getMessage(), BaseElement.NAME_ERROR);
     }
